requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
GitPython==3.1.37
requests==2.31.0
bcrypt==4.0.1
redis==5.0.1
pyvmomi==8.0.2.0
schedule==1.2.1